# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cartly vendor marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: rating
          in: query
          schema:
            type: number
        - name: availableOn
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorListResponse"
...
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendors/me:
    get:
      operationId: getMyVendorProfile
      tags: [vendors]
      summary: Get the authenticated vendor's own profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMyVendorProfile
      tags: [vendors]
      summary: Update the authenticated vendor's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetailResponse"
...
  /onboarding:
    post:
      operationId: completeOnboarding
      tags: [vendors]
      summary: Complete vendor onboarding (multi-step)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingBody"
      responses:
        "201":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetailResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    VendorSummary:
      type: object
      properties:
        id:
          type: number
        cartName:
          type: string
        category:
          type: string
        city:
          type: string
        coverPhoto:
          type: string
        startingPrice:
          type: number
        avgRating:
          type: number
        isActive:
          type: boolean
        subscriptionStatus:
          type: string
          enum: [active, inactive]
        onboardingComplete:
          type: boolean
      required: [id, cartName, category, city, isActive, subscriptionStatus, onboardingComplete]
    VendorDetail:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        cartName:
          type: string
        category:
          type: string
        bio:
          type: string
        city:
          type: string
        coverPhoto:
          type: string
        galleryPhotos:
          type: array
          items:
            type: string
        startingPrice:
          type: number
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PricingPackage"
        isActive:
          type: boolean
        subscriptionStatus:
          type: string
          enum: [active, inactive]
        avgRating:
          type: number
        profileViews:
          type: number
        totalReviews:
          type: number
        onboardingComplete:
          type: boolean
        isSaved:
          type: boolean
      required: [id, userId, cartName, category, city, isActive, subscriptionStatus, avgRating, profileViews, totalReviews, onboardingComplete]
    PricingPackage:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        description:
          type: string
      required: [name, price, description]
    VendorListResponse:
      type: object
      properties:
        vendors:
          type: array
          items:
            $ref: "#/components/schemas/VendorSummary"
        total:
          type: number
      required: [vendors, total]
    VendorDetailResponse:
      type: object
      properties:
        vendor:
          $ref: "#/components/schemas/VendorDetail"
      required: [vendor]
    UpdateVendorBody:
      type: object
      properties:
        cartName:
          type: string
        category:
          type: string
        bio:
          type: string
        city:
          type: string
        coverPhoto:
          type: string
        galleryPhotos:
          type: array
          items:
            type: string
        startingPrice:
          type: number
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PricingPackage"
        isActive:
          type: boolean
...
    OnboardingBody:
      type: object
      properties:
        cartName:
          type: string
        category:
          type: string
        bio:
          type: string
        city:
          type: string
        coverPhoto:
          type: string
        galleryPhotos:
          type: array
          items:
            type: string
        startingPrice:
          type: number
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PricingPackage"
        activateSubscription:
          type: boolean
      required: [cartName, category, bio, city, startingPrice]