# lib/api-zod/src/generated/types/onboardingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cartly vendor marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { PricingPackage } from "./pricingPackage";

export interface OnboardingBody {
  cartName: string;
  category: string;
  bio: string;
  city: string;
  coverPhoto?: string;
  galleryPhotos?: string[];
  startingPrice: number;
  packages?: PricingPackage[];
  activateSubscription?: boolean;
}